♻️ rename some components
