+++ eslint.config.js
@@ -21,7 +21,6 @@ export default [
       },
     },
     rules: {
-      "vue/multi-word-component-names": "off",
     },
   },
   {
