+++ src/components/ActionBarCollapseButton.vue
